+++ Sudokew/Sudokew.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29411.108
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudokew", "Sudokew\Sudokew.csproj", "{B25EDC9A-D184-4535-BAE8-A8953FFE1B19}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B25EDC9A-D184-4535-BAE8-A8953FFE1B19}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B25EDC9A-D184-4535-BAE8-A8953FFE1B19}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B25EDC9A-D184-4535-BAE8-A8953FFE1B19}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B25EDC9A-D184-4535-BAE8-A8953FFE1B19}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {CD1991AC-803F-477B-B8FC-6FEA789FB77D}
+	EndGlobalSection
+EndGlobal
