--- WindowsFormsApp5/WindowsFormsApp5.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29411.108
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsFormsApp5", "WindowsFormsApp5\WindowsFormsApp5.csproj", "{132CA375-1995-4E6B-9897-D693710486CB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{132CA375-1995-4E6B-9897-D693710486CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{132CA375-1995-4E6B-9897-D693710486CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{132CA375-1995-4E6B-9897-D693710486CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{132CA375-1995-4E6B-9897-D693710486CB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6092EF96-CD07-47F9-8A96-39A46592EC77}
-	EndGlobalSection
-EndGlobal
